ThreadSafeDicts.jl

A thread-safe Dict type for Julia programming

<br> <img src="https://github.com/wherrera10/ThreadSafeDicts.jl/blob/master/docs/src/spool.png"> <br><br>

Struct and Constructors

ThreadSafeDict

struct ThreadSafeDict{K,V} <: AbstractDict{K,V}
    dlock::ReentrantLock
    d::Dict{K,V}
    ThreadSafeDict{K,V}() where {K,V} = new(ReentrantLock(), Dict{K,V}())
    ThreadSafeDict{K,V}(d::Dict{K,V}) where {K,V} =
        new(ReentrantLock(), copy(d))
    ThreadSafeDict{K,V}(itr) where {K,V} =
        new(ReentrantLock(), Dict{K,V}(itr))
end

ThreadSafeDict()
ThreadSafeDict(d::Dict{K,V}) where {K,V}
ThreadSafeDict(itr)

Struct and constructors for ThreadSafeDict. There is one private ReentrantLock per Dict struct. All functions lock this lock, pass arguments to the d member Dict, unlock the ReentrantLock, and then return what is returned by the Dict.

Installation

using Pkg
Pkg.add("ThreadSafeDicts")

Functions Reference

This block automatically extracts docstrings from your source code:

ThreadSafeDicts.ThreadSafeDict

Base.delete!

Base.empty!

Base.get

Base.get

Base.get!

Base.get!

Base.getindex

Base.haskey

Base.iterate

Base.iterate

Base.length

Base.pop!

Base.print

Base.setindex!

ThreadSafeDicts.ThreadSafeDict — Type

ThreadSafeDict(pairs::Vector{Pair{K,V}})

Struct and constructor for ThreadSafeDict. There is one lock per Dict struct. All functions lock this lock, pass arguments to the d member Dict, unlock the spinlock, and then return what is returned by the Dict.

Base.delete! — Method

delete!(dic::ThreadSafeDict, k)

delete key k and its value from the dict

Base.empty! — Method

empty!(dic::ThreadSafeDict)

Remove all keys and values from the Dict

Base.get! — Method

get!(f::Function, dic::ThreadSafeDict, k)

Get value at key k if exists, otherwise set value at k to f() and return that value.

Base.get! — Method

get!(dic::ThreadSafeDict, k, v)

Get value at key k if exists, otherwise set value at k to v and return v.

Base.get — Method

get(f::Function, dic::ThreadSafeDict, k)

Get value at key k if exists, otherwise return f()

Base.get — Method

get(dic::ThreadSafeDict, k, v)

Get value at key k if exists, otherwise return v

Base.getindex — Method

getindex(dic::ThreadSafeDict, k)

Get the value at key index k.

Base.haskey — Method

haskey(dic::ThreadSafeDict, k)

Return true if key k is in the dict, else return false.

Base.iterate — Method

iterate(dic::ThreadSafeDict, i)

Iterate through the Dict returning its key-value pairs.

Base.iterate — Method

iterate(dic::ThreadSafeDict)

Iterate through the Dict returning its key-value pairs.

Base.length — Method

length(dic::ThreadSafeDict)

Return the length of the Dict, considered as a vector of key-value pairs

Base.pop! — Method

pop!(dic::ThreadSafeDict)

remove and return a key-value pair from the Dict

Base.print — Method

print(io::IO, dic::ThreadSafeDict)

Print the ThreadSafeDict, including the state of its lock and contents of the undelying Dict.

Base.setindex! — Method

setindex!(dic::ThreadSafeDict, k, v)

Set the value at key index k to v.